Moved log to its own header and prevented multiple log init
